added updater packaging, added params for skipping

+++ build.ps1
@@ -1,5 +1,8 @@
 param (
-    [switch] $SkipBuild
+    [switch] $SkipMainBuild,
+    [switch] $SkipMainApp,
+    [switch] $SkipUpdaterBuild,
+    [switch] $SkipUpdaterApp
 )
 
 $outFolder = "entrabulker"
@@ -13,7 +16,7 @@ $mainReact = "01-bulker-app"
 $updaterReact = "02-updater-app"
 
 # main app packaging
-if(!($skipBuild)){
+if(!($SkipMainBuild)){
     cd "$mainReact"
     npm run build
 
@@ -21,11 +24,28 @@ if(!($skipBuild)){
 
     mv "$mainReact\dist" ".\$outFolder\apps\madist" # madist is the main dist folder name
 }
+if(!($SkipMainApp)){
+    $fileName = "EntraBulker"
+    pyinstaller --onefile --noconsole --name "$fileName" ".\backend\main.py"
 
-$fileName = "EntraBulker"
-pyinstaller --onefile --noconsole --name "$fileName" ".\backend\main.py"
-
-mv ".\dist\$fileName.exe" ".\$outFolder\apps\$fileName.exe" -force
-rmdir ".\dist" -ea 0
+    mv ".\dist\$fileName.exe" ".\$outFolder\apps\$fileName.exe" -force
+    rmdir ".\dist" -ea 0
+}
 
 # updater app packaging
+# all updater contents will be in the parent folder from the root (apps)
+if(!($SkipUpdaterBuild)){
+    cd "$updaterReact"
+    npm run build
+
+    cd "$projRoot"
+    mv "$updaterReact\dist" ".\$outFolder\udist" # udist is the updater dist folder name
+}
+
+if(!($SkipUpdaterApp)){
+    $fileName = "EntraUpdater"
+    pyinstaller --onefile --noconsole --name "$fileName" ".\backend\updater_main.py"
+
+    mv ".\dist\$fileName.exe" ".\$outFolder\$fileName.exe" -force
+    rmdir ".\dist" -ea 0
+}
